services/gateway/pyproject.toml
[project]
name = "gateway"
version = "0.1.0"
description = "API gateway for artwork matching"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
    "httpx>=0.28.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=6.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "pyright>=1.1.390",
    "bandit>=1.7.0",
    "deptry>=0.21.0",
    "codespell>=2.3.0",
    "semgrep>=1.99.0",
    "pip-audit>=2.7.0",
    "pygount>=1.8.0",
    "types-PyYAML>=6.0.0",
    "pillow>=10.0.0",
    "respx>=0.21.0",
    "numpy>=1.26.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/gateway"]

# === Ruff ===
[tool.ruff]
target-version = "py312"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E", "W", "F", "I", "B", "C4", "UP", "ARG", "SIM", "TCH", "PTH", "ERA", "PL", "RUF",
]
ignore = ["PLR0913", "PLR0915", "PLR2004"]

[tool.ruff.lint.isort]
known-first-party = ["gateway"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

# === Mypy ===
[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_configs = true
show_error_codes = true
files = ["src"]

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false

# === Bandit ===
[tool.bandit]
exclude_dirs = ["tests", ".venv"]
skips = ["B101"]

# === Pytest ===
[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
asyncio_mode = "auto"
addopts = "-v --tb=short"
markers = [
    "unit: Unit tests (fast, isolated with mocks)",
    "integration: Integration tests (real HTTP calls, mocked backends)",
    "performance: Performance benchmark tests",
    "slow: Slow tests (excluded from default run)",
]

# === Coverage ===
[tool.coverage.run]
source = ["src"]
branch = true
omit = ["tests/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover", "def __repr__", "raise NotImplementedError",
    "if TYPE_CHECKING:", "if __name__ == .__main__.:",
]
fail_under = 80
show_missing = true

# === Deptry ===
[tool.deptry]
extend_exclude = [".venv", "tests"]

[tool.deptry.per_rule_ignores]
DEP002 = [
    "pydantic-settings",  # Used via pydantic imports
    "pytest",             # CLI tool
    "pytest-cov",         # pytest plugin
    "pytest-asyncio",     # pytest plugin
    "ruff",               # CLI tool
    "mypy",               # CLI tool
    "pyright",            # CLI tool
    "bandit",             # CLI tool
    "deptry",             # CLI tool
    "codespell",          # CLI tool
    "semgrep",            # CLI tool
    "pip-audit",          # CLI tool
    "pygount",            # CLI tool
    "types-PyYAML",       # Type stubs
    "pillow",             # Used in test factories
    "respx",              # Used in integration tests
    "numpy",              # Used in performance test generators
]
